+++ packages/cubejs-query-orchestrator/src/orchestrator/QueryOrchestrator.ts
@@ -30,6 +30,26 @@ export interface QueryOrchestratorOptions {
   skipExternalCacheAndQueue?: boolean;
 }
 
+function detectQueueAndCacheDriver(options: QueryOrchestratorOptions): CacheAndQueryDriverType {
+  if (options.cacheAndQueueDriver) {
+    return options.cacheAndQueueDriver;
+  }
+
+  if (getEnv('cacheAndQueueDriver')) {
+    return getEnv('cacheAndQueueDriver');
+  }
+
+  if (getEnv('redisUrl') || getEnv('redisUseIORedis')) {
+    return 'redis';
+  }
+
+  if (getEnv('nodeEnv') === 'production') {
+    return 'cubestore';
+  }
+
+  return 'memory';
+}
+
 export class QueryOrchestrator {
   protected readonly queryCache: QueryCache;
 
@@ -50,30 +70,31 @@ export class QueryOrchestrator {
     options: QueryOrchestratorOptions = {}
   ) {
     this.rollupOnlyMode = options.rollupOnlyMode;
-
-    const cacheAndQueueDriver = options.cacheAndQueueDriver || getEnv('cacheAndQueueDriver') || (
-      (getEnv('nodeEnv') === 'production' || getEnv('redisUrl') || getEnv('redisUseIORedis'))
-        ? 'redis'
-        : 'memory'
-    );
-    this.cacheAndQueueDriver = cacheAndQueueDriver;
+    const cacheAndQueueDriver = detectQueueAndCacheDriver(options);
 
     if (!['redis', 'memory', 'cubestore'].includes(cacheAndQueueDriver)) {
-      throw new Error('Only \'redis\', \'memory\' or \'cubestore\' are supported for cacheAndQueueDriver option');
+      throw new Error(
+        `Only 'cubestore', 'redis' or 'memory' are supported for cacheAndQueueDriver option, passed: ${cacheAndQueueDriver}`
+      );
     }
 
     const { externalDriverFactory, continueWaitTimeout, skipExternalCacheAndQueue } = options;
 
     const redisPool = cacheAndQueueDriver === 'redis' ? new RedisPool(options.redisPoolOptions) : undefined;
     this.redisPool = redisPool;
+    this.cacheAndQueueDriver = cacheAndQueueDriver;
 
     const cubeStoreDriverFactory = cacheAndQueueDriver === 'cubestore' ? async () => {
-      const externalDriver = await externalDriverFactory();
-      if (externalDriver instanceof CubeStoreDriver) {
-        return externalDriver;
+      if (externalDriverFactory) {
+        const externalDriver = await externalDriverFactory();
+        if (externalDriver instanceof CubeStoreDriver) {
+          return externalDriver;
+        }
+
+        throw new Error('It`s not possible to use Cube Store as queue/cache driver without using it as external');
       }
 
-      throw new Error('It`s not possible to use CubeStore as queue & cache driver without using it as external');
+      throw new Error('Cube Store was specified as queue/cache driver. Please set CUBEJS_CUBESTORE_HOST and CUBEJS_CUBESTORE_PORT variables. Please see https://cube.dev/docs/deployment/production-checklist#set-up-cube-store to learn more.');
     } : undefined;
 
     this.queryCache = new QueryCache(